finished login / register routes
