+++ src/__init__.py
@@ -31,8 +31,8 @@ def create_app():
 
     db.init_app(app)
 
-    from .public import public
+    from .auth import auth
 
-    app.register_blueprint(public)
+    app.register_blueprint(auth)
 
     return app
