+++ src/dump.py
@@ -1,0 +1,8 @@
+import json
+from aiohttp import web
+
+
+def json_response(body="", **kwargs):
+    kwargs["body"] = json.dumps(body or kwargs["body"]).encode("utf-8")
+    kwargs["content_type"] = "text/json"
+    return kwargs["body"]
